spef_extractor: remove dependency on matplotlib and numpy
Even if these are widely used libraries, they are not needed for
spef extraction.
