+++ scripts/spef_extractor/lef_def_parser/lef_parser.py
@@ -106,40 +106,6 @@ class LefParser:
         print ("Parsing LEF file done.")
 
 
-def draw_cells():
-    """
-    code to draw cells based on LEF information.
-    :return: void
-    """
-    to_draw = []
-    to_draw.append(input("Enter the first macro: "))
-    to_draw.append(input("Enter the second macro: "))
-    #to_draw = ["AND2X1", "AND2X2"]
-
-
-    plt.figure(figsize=(12, 9), dpi=80)
-    plt.axes()
-
-    num_plot = 1
-    for macro_name in to_draw:
-        # check user's input
-        if macro_name not in lef_parser.macro_dict:
-            print ("Error: This macro does not exist in the parsed library.")
-            quit()
-        macro = lef_parser.macro_dict[macro_name]
-        sub = plt.subplot(1, 2, num_plot)
-        # need to add title
-        sub.set_title(macro.name)
-        draw_macro(macro)
-        num_plot += 1
-        # scale the axis of the subplot
-        plt.axis('scaled')
-
-
-    # start drawing
-    print ("Start drawing...")
-    plt.show()
-
 """
 # Main Class
 if __name__ == '__main__':
